Resolve sibling projects directly.
That way we know their metadata without them first having to be added to the
workspace, or known to the Pacman repository that launched Scaled.

This makes it easier to work on a project that uses the Scaled build system but
which is otherwise not associated with your Scaled installation.